:necktie: Separate out admin toolbar, move js into components
